Split out thirdsOverlay code to separate file
